% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdt-utils.R
\name{tdt_format_random_effects}
\alias{tdt_format_random_effects}
\title{Format bare names as \code{(1 | name)} random-effect terms}
\usage{
tdt_format_random_effects(random_effects = NULL)
}
\description{
Format bare names as \code{(1 | name)} random-effect terms
}
\keyword{internal}
