% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdt-utils.R
\name{tdt_quantile}
\alias{tdt_quantile}
\title{Quantile wrapper with TDT-friendly defaults}
\usage{
tdt_quantile(x, probs = c(0.025, 0.5, 0.975))
}
\arguments{
\item{x}{Numeric vector.}

\item{probs}{Numeric vector of quantile probabilities.}
}
\value{
Numeric vector of length \code{length(probs)}.
}
\description{
Quantile wrapper with TDT-friendly defaults
}
\examples{
tdt_quantile(rnorm(100))
}
