% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdt-utils.R
\name{tdt_random_effect_variables}
\alias{tdt_random_effect_variables}
\title{Extract variable names from random-effect terms}
\usage{
tdt_random_effect_variables(random_effects = NULL)
}
\description{
Extract variable names from random-effect terms
}
\keyword{internal}
